work on users branches module

+++ src/modules/user_branches/userBranches.model.js
@@ -3,21 +3,39 @@ import db from "../../common/config/db.js";
 /*---------------- Assign User Branch ----------------*/
 
 const assignUserBranch = async (user_id, branch_id) => {
-  const [result] = await db.query(
+  /*-------------Insert data----------*/
+  await db.query(
     `
-    INSERT INTO user_branches
-    (
-      user_id,
-      branch_id
-    )
+    INSERT INTO user_branches (user_id, branch_id)
     VALUES (?, ?)
     `,
     [user_id, branch_id],
   );
 
-  return result;
-};
+  /*------------Fetch inserted data-------------------*/
 
+  const [rows] = await db.query(
+    `
+  SELECT
+    ub.user_id,
+    u.name AS user_name,
+    ub.branch_id,
+    b.name AS branch_name
+  FROM user_branches ub
+
+  JOIN users u
+    ON u.user_id = ub.user_id
+
+  JOIN branches b
+    ON b.branch_id = ub.branch_id
+
+  WHERE ub.user_id = ?
+  AND ub.branch_id = ?
+  `,
+    [user_id, branch_id],
+  );
+  return rows[0];
+};
 /*---------------- Get All User Branches ----------------*/
 
 const getAllUserBranches = async () => {
@@ -73,15 +91,40 @@ const getUserBranches = async (user_id) => {
 /*---------------- Delete User Branch ----------------*/
 
 const deleteUserBranch = async (id) => {
-  const [result] = await db.query(
+  /*------- Get branch data before delete-----------*/
+  const [rows] = await db.query(
     `
-      DELETE FROM user_branches
-      WHERE user_branch_id = ?
-      `,
+    SELECT
+      ub.branch_id,
+      b.name AS branch_name
+    FROM user_branches ub
+
+    JOIN branches b
+      ON b.branch_id = ub.branch_id
+
+    WHERE ub.user_branch_id = ?
+    `,
     [id],
   );
 
-  return result;
+  /*------- Check if record exists-------------*/
+  if (rows.length === 0) {
+    throw new Error("User branch not found");
+  }
+
+  /*---------- Store deleted data------------*/
+  const deletedData = rows[0];
+
+  /*--------- Delete record-------------------*/
+  await db.query(
+    `
+    DELETE FROM user_branches
+    WHERE user_branch_id = ?
+    `,
+    [id],
+  );
+
+  return deletedData;
 };
 
 export default {
